backend/pyproject.toml
[tool.poetry]
name = "research_canvas"
version = "0.1.0"
description = "To run backend services"
authors = ["dharun-karthick-ramaraj <dharun.ramaraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
boto3 = "^1.35.57"
pandas = "^2.2.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
snowflake-connector-python = "^3.12.3"
pinecone-client = "^5.0.1"
langchain-nvidia-ai-endpoints = "^0.3.5"
langchain-pinecone = "^0.2.0"
langgraph = "^0.2.45"
copilotkit = "^0.1.27"
langchain-core = "^0.3.15"
langchain-openai = "^0.2.6"
langchain-community = "^0.3.5"
langchain-anthropic = "^0.2.4"
langchain-google-genai = "^2.0.4"
langchain = "^0.3.7"
openai = "^1.54.3"
tavily-python = "^0.5.0"
uvicorn = "^0.32.0"
html2text = "^2024.2.26"
fastapi = "^0.115.5"
typing-extensions = "^4.12.2"
arxiv = "^2.1.3"
pdfkit = "^1.0.0"
markdown2 = "^2.5.1"
wkhtmltopdf = "^0.2"
google-auth = "^2.36.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.153.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
demo = "research_canvas.demo:main"